backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pydantic
pydantic-settings
python-dotenv
langchain
langchain-openai
langchain-community
langchain-core
langchain-experimental
langchain-classic
pandas
scikit-learn
redis
python-multipart
pypdf
openpyxl
langchain-ollama
bcrypt
python-jose[cryptography]
httpx
